fix: change `last_updated` column type from `BIGINT` to `INTEGER` in user and group profiles tables

+++ migrations/20260212174800_create_profiles_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS user_profiles (
     user_id TEXT PRIMARY KEY,
     encrypted_blob BLOB NOT NULL,
-    last_updated BIGINT NOT NULL,
+    last_updated INTEGER NOT NULL,
     version INTEGER DEFAULT 1
 );
